Refactor provider to own file
