+++ src/App.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import { LightboxProvider } from './hooks/useLightbox';
+import { LightboxProvider } from './hooks/LightboxProvider';
 import './index.css';
 import AppRoutes from './routes';
